+++ grpcplugin/grpc.go
@@ -14,12 +14,13 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-package plugin
+// Package grpcplugin is a gRPC based plugin system.
+package grpcplugin
 
 import (
 	"fmt"
 
-	"github.com/katzenpost/server/plugin/proto"
+	"github.com/katzenpost/server/grpcplugin/proto"
 	"golang.org/x/net/context"
 )
